PSAD_MKDAD/requirements.txt
certifi==2021.10.8
cycler==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.3.0
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264701743/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.4.0.46
pandas==1.2.5
Pillow @ file:///tmp/build/80754af9/pillow_1625655818400/work
pyparsing==2.4.7
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1633452062248/work
PyYAML==5.3.1
scikit-learn==1.0
scipy==1.5.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
threadpoolctl==3.0.0
torch==1.8.2
torchaudio==0.8.2
torchvision==0.9.2
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1632313171031/work
